Input and Design Changes
Added input types and added some design changes. Changed the positioning of the overall card and the hyper link for the forgot password. Added padding to the content element so it's not flat with the edge of a browser window. Other minor changes made to layout, and moved tabbar to top of page.