Scroll to user (#326)
* Functioning scroll to specific point

Function added that allows scrolling to specific point in the leaderboard.

* Scrolls to user

Scrolls to the user when the buttons are pushed. However, it is not using updates states on click. The goal is to scroll to user on page load and on button press.

* Update LeaderBoardChart.tsx

Scrolling after double clicking team or individual.

* Update LeaderBoardChart.tsx

Call changes

* Scroll to User

Leaderboard now scrolls to the user so that it is easy for the user to find themselves.

* Scroll distance adjustment